Wetube Clone - File upload, DB Relationships, etc.

+++ wetube-clone/src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 
 export const home = async (req, res) => {
@@ -25,26 +26,35 @@ export const getUpload = (req, res) => {
 }
 
 export const postUpload = async (req, res) => {
+  const { user: { _id } } = req.session;
+  const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
+      fileUrl,
       title,
       description,
       hashtags: Video.formatHashtags(hashtags),
+      owner: _id,
     });
+
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    await user.save();
     return res.redirect("/");
   } catch(error) {
     console.log(error);
     return res.status(400).render("upload", { 
       pageTitle: "Upload Video", 
-      errorMessage: error._message,
+      errorMessage: error._message, 
     });
   }
 }
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
+  console.log(video);
 
   if (video === null) {
     return res.status(404).render("404", { pageTitle: "404 Not Found" });
@@ -54,21 +64,34 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id }
+  } = req.session; 
   const video = await Video.findById(id);
 
-  if(!video) {
+  if (!video) {
     return res.status(404).render("404", { pageTitle: "404 Not Found" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   return res.render("edit", { pageTitle: `Edit <${video.title}>`, video });
 }
 
 export const postEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id }
+  } = req.session; 
   const { title, description, hashtags } = req.body;
+  const video = Video.findById(id);
 
-  const isVideoExists = Video.exists({ _id: id });
-  if (!isVideoExists) {
+  if (!video) {
     return res.status(404).render("404", { pageTitle: "404 Not Found" });
+  }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
 
   await Video.findByIdAndUpdate(id, {
@@ -81,6 +104,22 @@ export const postEdit = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id }
+  } = req.session; 
+  const video = await Video.findById(id);
+  const user = await User.findById(_id);
+
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "404 Not Found" });
+  }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+
   await Video.findByIdAndDelete(id);
+  user.videos.splice(user.videos.indexOf(id), 1);
+  await user.save();
   return res.redirect("/");
 }
